+++ docs/src/history.md
@@ -1,6 +1,11 @@
 # Version history
 
 
+## Current development version
+
+* CHANGED: `getindex()` changed to `get()`, since we need a default value to be provided.
+
+
 ## v0.1.0
 
 * ADDED: more method variations to lift unwrapped values